feat: complete rework of all types for ReScript 10.1
The types of (React-)components have been reworked completely for the new JSX v4 syntax.
There are two different sets of bindings required to support jsx v3 and v4. By opening "SolidV3",
support for the older bindings can be restored.

BREAKING CHANGE: The bindings support the new ReScript (JSX v4) types by default.
In oder to keep the old behaviour the option `-open SolidV3` has to be added to the `bsc-flags` inside
the `bsconfig.json` configuration. This is analogous to the behaviour of the
current rescript-react